Remove environments from GitHub actions
GitHub environments are helpful for a few things:

- they let us group secrets by environment
- they give us a quick and easy view of what is deployed where
- they maintain a history of deployments

Our actions are not compatible with this, the deploy actions runs on
merges into main regardless [1] and includes logic inside the action -
this means the environments show all merges to main as being deployed,
when they are not.

We also predix our secrets with the environment so do not use the secret
grouping.

For this reason, we remove the entries to the GitHub environments and
delete the environments in the GitHub UI once this change is merged.

[1] https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/blob/393c8c9713c64c098d3494b4573ddcffa1b7160d/.github/workflows/build-and-push-image-test.yml#L6